requirements.txt
selenium~=4.38.0
pytest~=8.4.2
allure-python-commons~=2.15.0
webdriver-manager~=4.0.2